Improve behavior of cpu_is_imx6* macros
If the CONFIG_SOC_MX6* for some CPU type is not defined, the
corresponding cpu_is_mx6() function is still evaluated and causes
code for wrong CPU types to be included. This makes optimization of
the kernel for a reduced i.MX6 set difficult to impossible.

For example the file arch/arm/mach-imx/pm-imx6.c contains the following
section:

	if (cpu_is_imx6sl())
		imx6sl_set_wait_clk(false);

But if CONFIG_SOC_MX6SL is not set, the test for imx6sl is still done
and function imx6sl_set_wait_clk() must be available. But in this case
the file arch/arm/mach-imx/clk-imx6sl.c which contains this function
is excluded from the build. Therefore this will result in undefined
symbols when compiling the kernel.

This modification converts function cpu_is_imx6sl to a macro that is
replaced by 0 (false). So now the above if condition can be evaluated
to false at compile time and the compiler will remove the call to
imx6sl_set_wait_clk() completely.