Add support for SPI (spidev) on efusA9X
It depends on the presence of RTS on UAT_C if we can use one or two
chip selects on SPI_B. Therefore define a CONFIG_EFUSA9X_UART_C_RTS
that tells whether to use RTS on UART_C or not. This can be converted
to a Kconfig switch in the future.

Device tree modifications:

- Add pinctrl groups for ecspi1 and ecspi5
- Enhance and activate ecspi1 and ecspi5 entries, add an spidev entry
  for every available chips select signal
- Modify UART_C settings according to CONFIG_EFUSA9X_UART_C_RTS
- Remove QSPI settings

Other modifications:

- Add compatible entry "linux,spidev" for drivers/spi/spidev.c; the
  existing "rohm,dh2228fv" was really strange
- Add CONFIG_SPI_SPIDEV to fsimx6sx_defconfig