Add support for Atmel MXT touch for efusA9X
This driver needs a configuration file. The easiest way to create this
file is by using Atmel's driver version from

  https://github.com/atmel-maxtouch

This allows interactive modification of the controller objects by
using a separate application that can be compiled for ARM. This
application can be cloned from

  https://github.com/atmel-maxtouch/mxt-app.git

When the final setting is complete, the config can be stored as a file
*.cfg. This file has to be put into directory

  /lib/firmware

on the target. See /sys/bus/i2c/devices/?-004a/ for possible runtime
settings and also see

  mxt-app --help

for the features of this application.

Replacing the mainline Linux driver with Atmel's version affects the
following files:

- drivers/input/touchscreen/atmel_mxt_ts.c gets new version
- include/linux/i2c/atmel_mxt_ts.h gets replaced by new files
  include/linux/platform_data/atmel_mxt_ts.h and
  include/dt-bindings/input/atmel_mxt_ts.h
- New file Documentation/devicetree/bindings/input/amtel,maxtouch.txt

Device tree modifications:

- Activate mxt224 entry in I2C2 device list and add missing parameters

Other changes:

- In fsimx6sx_defconfig configure CONFIG_TOUCHSCREEN_ATMEL_MXT as
  module, otherwise it will not find its configuration file
  (searched like a firmware); start with modprobe atmel_mxt_ts
- Prepare an mxt224.cfg file to be used with MXT224 touch controller
  and WVGA display as available in F&S SKITs; this file is kept
  separate from the GIT repository at the moment