Add support for extended MXS DMA (APBH)
This modifed DMA engine is used in the F&S implementation of the
GPMI NAND driver. This version uses a ring buffer of DMA descriptors
and allows setting the data of the descriptor in a more simple way by
misusing the struct scatterlist pointer argument of function
mxs_dma_ext_prep_slave_sg() as a pinter to struct mxs_dma_ccw. This
allows the caller to prepare all necessary data.

- Add to drivers/dma/Kconfig and drivers/dma/Makefile
- Add driver file drivers/dma/mxs-dma-ext.c
- Add header file include/linux/mxs-dma-ext.h

Remark:
The purpose of using this DMA for a cyclic LCD framebuffer refresh
like in the original driver is not supported in this version.