fix gc_hal_kernel_hardware.c
from 14aug, delivery-patches.tgz,
        you derive these patches by comparing the
        l5.0.0_1.0.0-ga (Android) tag to the
        rel_imx_3.10.53_1.1.1 (regular non-Android) tag

compare commit from freescale:

MGS-608 [#1592] disable GC2000 TX clock gating

Before this change, the boot did show the boot animation,
but failed the show the regular Android UI. In the logcat,
it could be observed that the processes that were supposed to
present graphical output failed with errors akin to:
	I/OpenGLRenderer( 2826): Initialized EGL, version 1.4
	W/OpenGLRenderer( 2826): Failed to choose config with EGL_SWAP_BEHAVIOR_PRESERVED, retrying without...
	F/OpenGLRenderer( 2826): Failed to choose config, error = EGL_SUCCESS
and subsequent abort. The performance of the bootanimation
was very low.

With this change, the Android system will boot past
the boot animation into the user-operable GUI, though
with still very low graphics performance.

Signed-off-by: Simon Braunschmidt <sb@emlix.com>