Fix F&S gpmi-nand-fus.c driver
When reading data with ECC from flash, the driver assumed that the BCH
was always taking longer than the DMA and therefore only checked for
the BCH interrupt in these cases.

But this is not true. In fact with only very rare exceptions, the DMA
finishes *after* the BCH. So if the driver started a new DMA transfer
very quickly after receiving the BCH interrupt, it could have
accidentally caught the DMA interrupt from the previous DMA transfer
and immediately finished the current transfer again. This
resulted in sporadic "DMA chain execution failed" errors.

Now the driver checks for both interrupts in this case. This makes
sure that there are no pending interrupts anymore which might be
misinterpreted.