Add support for different touches on efusA7UL
Device tree modifications:

Add several touch devices for I2C_A:

- On-board 4/5 wire touch controller SX8655 (I2C address 0x48)
- 4/5 wire touch controller SX8655 on RGB adapter (I2C address 0x49)
- PCAP touch controller FT5x06 for touch connector (I2C address 0x38)
- PCAP touch controller MXT224 for touch connector (I2C address 0x4A)

Because all these touches use the same IRQ line, only one of these
touches can be active at a time. This can be set at the top of the
device tree by setting one of the CONFIG entries. This will be moved
to a Kconfig file or some similar infrastructure in the future.

By default only the SX8655 on the RGB adapter is active because this
is the one that we need for our SKIT with display.

Other modifications:

We also only activate the SX8655 driver in fsimx6ul_defconfig.