Add vmmc regulators to device trees of some F&S boards
Even if we do not have switched power supplies for MMC/SD, give some
fixed-voltage regulators. This makes it easier to extend the device tree
in case there will be a switched power supply in the future. In addition
it removes a warning in the boot messages.

efusA7UL:

- Add reg_wlan_vmmc_1v8 for WLAN
- Add reg_vref_3v3 as VMMC for SD cards and eMMC

efusA9X:

- Add reg_wlan_vmmc_1v8 for WLAN
- Use reg_vref_3v3 for SD cards and eMMC

BemA9X:

- Add reg_wlan_vmmc_1v8 for WLAN

In addition remove "keep-power-in-suspend" and "enable-sdio-wakeup"
from all MMC devices but WLAN chips.