Add mono2both property to SGTL5000 sound codec
When property "mono2both" is given in the I2C node of the SGTL5000 codec,
play mono sound on both channels instead of only the left channel.

Example:

&i2c1 {
	...
	sgtl5000: sgtl5000@0a {
		compatible = "fsl,sgtl5000";
		reg = <0x0a>;
		mono2both;
		clocks = <&clks ...>;
		VDDA-supply = <&reg_sgtl5000_vdda>;
		VDDIO-supply = <&reg_sgtl5000_vddio>;
		VDDD-supply = <&reg_sgtl5000_vddd>;
	};
	...
};

Add this property also to the device tree bindings documentation of
the sgtl5000.

Set mono2both as default on efusA9X and efusA7UL.

Remark: Using "mono2both" may consume slightly more power than just
using the left channel in case of mono. So it only makes sense if
the codec is actually connected with both channels (stereo).