remove + annotation appended to kernel name
If the kernel is build when git is not in a clean state a + annotation
is added to the kernel release name.
If the kernel is build in this state separately from buildroot this
leads to the problem that kernel modules cannot be loaded because
the kernel name in uImage differs from kernel name in rootfs.
This patch prevents the kernel from setting the + annotation.