Use F&S DDR3 settings
arch/arm/mach-imx/busfreq_ddr3.c changes DRAM setings on the fly depending
on the bus frequency. Unfortunately the initialization of the MMDC can
only read the register values of the MMDC itself, but not the DDR3
settings that were sent (in the bootloader) as commands to the RAM chips.
Therefore it needs to know some hard-coded values that must be the same
as in the bootloader. Modify the settings to reflect the F&S RAM settings,
do not use the fixed NXP settings anymore. This should make the whole
RAM system more stable.

FIXME:
On F&S boards, these values should be passed via NBootArgs to the kernel.
Using fix values bears the risk of inconsistent settings in bootloader
and kernel and also can not represent all possible combinations as they
depend on CPU type (D/Q or S/DL) and board (armStoneA9, QBlissA9, efusA9,
PicoMODA9).

For example we always activate On-Die-Termination (ODT), but it is only
required on armStoneA9 and PicoMODA9. So the RAM energy consumption is
unnecessary high on efusA9 and QBlissA9.