Improve armstonea9 DT pmic
The pmic is not correctly implemented yet, because the ldo-bypass for
arm and soc are now implemented in the official UBoot (before it was in
the kernel). We dont change the UBoot so we cant use the pmic. For the
future we must implement the ldo-bypass in the UBoot or change the code
in the Kernel to set in the PMU_REG_CORE register, core and arm power
domain to Power FET switched full on (11111). In the device tree i
implemented the devices for pmic support but commented it out till the
ldo-bypass (core and arm) will be implemented in the UBoot/Kernel code.