[ft5x06_ts_v2] Remove dubious update feature, rename to ft5x06_ts
Experience has shown that different displays with Focaltech touch
behave differently with regard to some settings like offset and gain.
This implies that they also have different firmwares. So the feature
to update the firmware was most probably not meant for all Focaltech
touches and could probably even disrupt the operation if applied to
a wrong device.

Therefore if it can not be used in a save way, remove the whole code
to avoid any negative implications. This also means that the file
ft_app.i can be removed, too.

Also rename the module back to ft5x06_ts.c to avoid the version number
in the file name. We are now at V3.0 anyway.