Fix fixup code for Micrel KSZ8081 PHY on efusA7UL
The fixup code in mach-imx6ul.c did not mask out the PHY chip revision,
which meant the fixpup for setting the registers was not run on a
revision > 0. Now use PHY_ID_KSZ8031 and MICREL_PHY_ID_MASK from
include/linux/micrel_phy.h.

The LED was also set to a wrong mode on efusA7UL. Fix this, too.