New display configuration for armStoneA9, armStoneA9r2, efusA9
Instead of defining up to four displays with settings for each possible
display type for each display, simply define one setting for each
display type and one display type for each mxcfb. This includes mode
string (for LCD and HDMI) and display timings (for LVDS) as well as
pixel format and default bits-per-pixel. So everything is in one place
at the beginning of the device tree file (q and dl file respectively).

Move all other display stuff to the qdl file for each platform.
Determining IPU and DI is also done automatically there.

MXCFB0 and MXCFB1 will get the video overlays if two IPUs are used,
If only one IPU is used, this is MXCFB0. This can be set by defining
CONFIG_*_USE_ONE_IPU on Dual/Quad. This is the default on Solo/DualLite).

Add comments to explain everything.

Remove LCD settings from armStoneA9r2, this is not supported there.