Add copy method for firmware files, use with mxt224.cfg
If a firmware file is not actually a binary, but a text file (for
example a kind of configuration), it is not necessary to encode it
with ihex or similar. Add a new rule to firmware/Makefile, that just
copies a file to the final name if it ends in ".copy".

Use this method for firmware/atmel/mxt224.cfg, which means renaming
the mxt224.cfg file to mxt224.cfg.copy.

Up tp now, the mxt224.cfg file was deleted if "make clean" or similar
targets were executed.