Improve FEC ethernet configuration on F&S boards
On efusA7UL, both PHYs are located on one MDIO bus. This can either
be the one from FEC1 or FEC2. We prefer FEC1 as long as FEC1 is
available.

In the previous code, this only worked if the MDIO bus was a child of
the FEC that was initialized first. Unfortunately FEC1 is located on
AIPS2 and FEC2 on AIPS1, which means that FEC2 was probed before FEC1.
So we should have added the MDIO bus to FEC2. And even more, the
interfaces were named in the order of registration, i.e. FEC2 was
eth0 and FEC1 was eth1. This was not how we wanted it to be.

Improve the FEC driver to handle this better.

- In fec_probe(), if a phy-handle is given, and the according MDIO
  bus is not yet available and the MDIO bus is not a child of this
  FEC port, then defer the probing until the MDIO bus is available.
- This obsoletes the code for MDIO bus sharing that was available in
  fec_enet_mii_init().
- Get the eth port number from the ethernet alias, not by increasing
  the number automatically.
- Do not count the MDIO bus occurences to know when to unregister
  the MDIO bus in fec_enet_mii_remove(). Set flag "have_own_mii_bus"
  instead that tells exactly whether this FEC uses the own MDIO bus
  or not.
- Activate second ethernet port in device trees of efusA7UL and
  efusA9X, add defines to allow for one, two or none ethernet ports.
- In mach-imx6ul.c, move functions imx6ul_opp_check_speed_grading()
  and imx6ul_opp_init() behind imx6ul_enet_init(), i.e. out of the
  region guarded by #ifdef CONFIG_FEC.