Improve armstonea9r2 DT audio sgtl5000 clock
Improve the clock chain of the sgtl 5000 codec. Until now the clock for
the sgtl5000 (cko1) and the clock of the mipi camera have the same clock
frequence because the output frequence were on cko1 and cko2. So if you
change the sgtl5000 clock you automatically change the mipi clock too
because they depend on each other. Now you can change the clock of the
sgtl5000 independent from the mipi clock.