Add Gamma2 option to leds-pwm.c driver
The human eye does not perceive the brightness of an LED as a linear
correlation to the PWM value. Usually a Gamma of 2.2 or 2.4 is used.
However such Gammas are difficult to compute in the kernel without
having floating point and a power function. But a Gamma of 2.0, which
is in fact a square function, does work well and is much better than
the linear mapping.

Add an option to the leds_pwm driver for a Gamma=2.0 (square) instead
of linear mapping.

- Change active_low entry of struct led_pwm to a flags value. Add flags
  LED_FLAGS_ACTIVE_LOW and LED_FLAGS_GAMMA_2 as predefined values in
  include/linux/leds_pwm.h.
- Do the same for struct led_pwm_data in drivers/leds/leds-pwm.c.
- Also look for binaty property gamma2 when parsing the device tree in
  function led_pwm_create_of().
- Use a different duty cycle computation in function led_pwm_set() if
  LED_FLAGS_GAMMA_2 is set.