Fix ethernet clock settings on i.MX6UL and i.MX6SX
When the ethernet clock is configured as output, the previous
implemantation to set some bits in gpr1 would only work if
regmap_update_bits() was realized as

  reg = (reg & ~mask) | val;

But it is actually imlemented as

  reg = (reg & ~mask) | (val & mask);

Therefore setting the mask to 0 does not work. Fix this by setting
also the mask to the bits to set.