Do not rotate display by default on aST-ADP-EDT1
The LCD display ET070080 shipped with F&S starter kits is connected
via the aST-ADP-EDT1 adapter to the efus baseboard or armStoneA9.
On this adapter, there is a PCA9632 LED driver chip that is used to
switch backlight power, dim the backlight brightness and switch the
display rotation signal. In the past this rotation signal was set to
high at start-up, which means the display was rotated by 180 degrees.

Modify all affected device trees to keep the rotate signal low by
default to have the native orientation of the display. This is
efusa7ul.dts, efusa9x.dts, efusa9qdl.dtsi and armstonea9qdl.dtsi.