Improve spi-imx driver support more than 4 cs
The default spi-imx driver only supports up to 4 cs. So we improve the
driver to use more than 4 chip selects.
Explanation:
CS 0 => 0000
CS 1 => 0001
CS 2 => 0010
CS 3 => 0011
CS 4 => 0100 that causes an error because we will write the value to the
wrong bits in the CPU register and the SPI hangs up. To avoid this we
use an AND gate to clear bit 2 and 3. After the change it works like
this:
CS 3 => 0011
CS 4 => 0100 & 0011 => 0000
CS 5 => 0101 & 0011 => 0001
...