Improve clock enable of imx6ul
Some clocks must be enabled at the startup of linux. So we implemented a
method to enable a clock at startup. For example the phy dp83848j needs
to communicate over mdio, on each phy an enabled reference clock.
Normally the driver enable the clock before it communicate with the phy
but if you have more than one phy on one mdio bus, then only the clock
will be enabled which gets initialized first. So the second clock will
not be enabled and the second phy will not be found as a device. This
causes that you cant bring up the phy because there is no device.