Improve auxiliary core driver
- if you echo stop to bootaux and after that echo start to bootaux then
you can transfer a binary to the tcm. That doesnt work before because
M4_Core was disabled by the echo stop and then you cant transfer any
image the tcm.
- if you doesnt have enabled the ccm clock then you cant transfer any
image to the tcm because the ccm clock is required for the tcm. We
have checked it but doesnt jump out of the function so now its fixed.