mxs-dma-ext.c: Add iram option, remove runtime suspend
Add option to use internal SRAM (iram) for DMA descriptors. This is done
similar to the SDMA code. If

  iram = <&ocram>;

is present in the dma_apbh device tree node, the descriptors are
allocated with gen_pool_dma_alloc() from the given iram pool (i.e.
ocram). Otherwise they are allocated with dma_alloc_coherent() from
the main memory pool.

To avoid wasting valuable SRAM space, reduce the DMA pool size from
four to one pages of PAGE_SIZE. This means only 113 instead of 455
descriptors. But the maxmimum number of descriptors used by GPMI NAND
driver is just 7 at the moment, so 113 still is way enough.

Remove the Runtime Power Manager functions mxs_dma_ext_runtime_suspend()
and mxs_dma_ext_runtime_resume(). They were never activated and they are
also removed in the original code of mxs_dma.c in the meantime. The code
was not working anymore anyway. Runtime power suspend would only save
power if the DMA clock could be deactivated, which is not possible in
the current layout of the driver.