Fix RTS/CTS settings for UART on fsimx6/sx/ul
The i.MX6 supports DTE and DCE mode for the UART devices. However the
pad names are all named after the DCE variant and unfortunately NXP has
mixed up the names for TX and RX, so instead of using TX for the
incoming data and RX for the outgoing data, as it would be correct for
a DCE device, they use TX for the outgoing and RX for the incoming data.
This is correct for a DTE device, but not a DCE device. So we have to
deal with this inconsistency.

So despite the fact that an F&S board is by definition a DTE device,
F&S uses the UART in DCE mode, where TX and RX names match with DTE.
But then RTS and CTS have to be connected in reverse, i.e. the RTS
signal (from the outside DTE view) needs to be connected to the CTS
pad of the SoC and is an output, and the outside CTS signal has to be
connected to the RTS pad of the SoC and is an input.

While these connections are valid in the hardware (at least in the
newest board revisions), this was not configured correctly in several
device trees. So the CTS pad actually has to be configured for CTS,
even if it is eventually our RTS signal, and the RTS pad has to be
configured for RTS, even if it is our CTS signal.

A port used for RS485, who just needs an RTS signal, actually needs the
CTS pad of the SoC and the CTS configuration for this pad.

CPU Pad   F&S Signal to outside
-------   ---------------------
TX     -->    TX (output)
RX     <--    RX (input)
CTS    -->    RTS (output)
RTS    <--    CTS (input)