Move SDMA descriptors to OCRAM on i.MX6 and i.MX6SX
For Suspend-To-RAM, the device descriptors of SDMA should be located in
internal OCRAM, not in DDR3 RAM, because DDR3 will be put to self refresh
mode while suspended. Like on i.MX6UL before, add property

  iram = <&ocram>;

to the sdma node in imx6qdl.dtsi and imx6sx.dtsi. This will locate
the descriptors in OCRAM when the SDMA driver is started.