SGTL5000 warmstart problems: restore code with internal LDO
Using the external VDDD instead of internal LDO causes problems on every
second warmstart (e.g. with the reset button) on F&S boards. The SGTL5000
codec is not detected in these cases.

Restore the code to use the internal LDO until this problem is further
analyzed and fixed.

The code to force the LO and HP pins always enabled should also only be used
in the case when property "no-standby" is given. So move these lines into
the appropriate if clause in sgtl5000_probe().

While we are at it, clean up the code, e.g. remove trailing whitespace.