sdhci-esdhc-imx.c: Do not switch pad settings unnecessarily often
When the device tree includes pad settings for Ultra High Speed (UHS) cards,
i.e. for 100MHz and 200MHz in addition to the default pad settings, then the
driver should not switch pad settings for each access. This is way too often
and unnecessary.

Add entry uhs to struct pltfm_imx_data that holds the current speed and only
change pad settings in esdhc_change_pinstate() if the speed changes.