sdhci.c: Remove console message if voltage switching fails
Convert messages of the kind

  "Switching to x.xV signalling voltage failed"

from warning messages to debug messages. They are good to detect if something
is configured correctly or not, but they are not necessary in a final product.

For example if a WLAN device supports different voltages for I/O, but the
board only supports 1.8V, then this is perfectly OK, but the driver always
tries 3.3V first and issues this message before it also tries 1.8V. The
device will work fine, so these messages are only an annoyance. This part
of the code is reworked in newer kernel versions anyway.