wm97xx-core.c: Remove warning about unspecified IRQ
Convert message

  "No IRQ specfied"

from an error message to a debug message. Not all use cases for this audio
codec involve touch support and then this message comes up. But this is no
error.

FIXME: In fact there should be a way in the device tree to specify if the
touch part of the device is used or not. Currently this is done by platform
support code (see drivers/input/touchscreen/{zylonite,mainstone}-wm97xx.c}
which is not appropriate anymore in the times of device trees.