Improve imx6sx can clock settings
The can_sel clock has only been enabled if the Cortex M4 was not
running, so if the Cortex-M4 was running and didn't enable the can clock
it didn't get enabled at all and the can driver couldn't be loaded.

Now the can_sel clock only gets endabled if the Cortex-M4 dind't enable
it before.