Improve auxiliary core
We have improved the usage of the auxiliary core. We have implement a
state machine to control the auxiliary core. Now we have 4 states off,
stopped, running, paused.
To reach this states you have the following commands:
    bootaux start
    bootaux stop
    bootaux pause
    bootaux continue
    bootaux stop
    bootaux run/"addr"
The Explanation of each state is documented in
Documentation/auxiliary_core.txt.
- We improve also that the M4 Clock prepare_enable/disable call. Now the
functions will be only called one time so if its already enabled by a
call before we dont increment the prepare_enable count anymore. The same
is done for prepare_disable.