Change fsimx6sx/ul_defconfig to support Silex WLAN
The WLAN driver for Silex SXSDPAC is built out-of-tree as an external kernel
module. However this external build process also creates an own sxcfg80211.ko
module which is a replacement for the builtin cfg80211.ko. Therefore we need
to configure CONFIG_CFG80211 as module. Also activating CONFIG_NL80211_TESTMODE
and CONFIG_CFG80211_WEXT is necessary. To have access point support, also add
CONFIG_HOSTAP.

We do not support the Taiyo Yuden WYSBCVGXA anymore that was available on
earlier board revisions of efusA9X and efusA7UL. Therefore we can remove
CONFIG_MWIFIEX.

On fsimx6sx we can also remove CONFIG_MAC80211. At the moment we do not have
a WLAN chip there that depends on the generic wireless stack. On fsimx6ul
however we need to support the Jorjin WG7801 WLAN chip for the Cube2.0 which
is based on WL18xx, so here we still need CONFIG_MAC80211.