Add phy led trigger
If a phy gets attached then there will be triggers created. For each
possible speed there will be one trigger registered. If the phy gets
detached the different triggers will be unregistered. If the speed
arrangement is completed the trigger with this specific speed is active.

The driver is backported from mainline kernel V4.10. (commit
2e0bc452f4721520502575362a9cd3c1248d2337)