Re-enable operating points for 792 and 900 MHz on i.MX6ULL
Even though DVS is not fully supported in Linux and the anatop LDO driver is
not yet modified, we can use these high frequencies now if DVS is disabled
and LDO is enabled. Disabling DVS is handled by a patch in U-Boot and the
current settings already have the anatop LDOs enabled.