Add property fsl,rx_fifo_trig to i.MX6 UART driver
The i.MX UART driver is optimized for large file transfers. This is achieved
by using the UART FIFO with trigger level 16. This means the software driver
is only triggered after at least 16 bytes have been received. If less data
is received, data is only passed on if a timeout expires. This may constitute
a problem if only low volume data is tranferred, for example a protocol with
single-byte commands. In this case the FIFO is never filled and reception
speed is limited by the timeout value. However if the user knows this, he
may desire to set the trigger level manually to a lower value, like 1 in
this case.

This commit adds an additional property fsl,rx_fifo_trig to do this. The
value is the number of bytes that must have been received to trigger the
interrupt or DMA request.

For example to set the FIFO trigger threshold of uart3 to 4 bytes, add
the following device tree node.

&uart3 {
	pinctrl-names = "default";
	pinctrl-0 = <&pinctrl_uart3>;
	fsl,rx_fifo_trig = <4>;          /* add this line */
	fsl,uart-has-rtscts;
	status = "okay";
};

If the property is not present, the driver behaves as before. The maximum
value is 32, at least on i.MX6 CPUs.