Add analog video output to fsimx6, using BT656+ADV7391
Add support for analog video output to all boards of the fsimx6 family that
have RGB-LCD output, i.e. efusA9, armStoneA9, NetDCUA9 and PicoCOMA9. Using
this feature requires an externally connected video encoder of the type
ADV7391. It is accessed via the low 8 data bits of the LCD port. It uses
BT656 protocol. In addition a I2C connection is required to set the registers
of the ADV7391.

Device tree modifications:

- Add option to use BT656 output as additional option to LCD, LVDS0/1, HDMI
- Add default settings for BT656
- Add bt656 and adv7391 device nodes
- Add IOMUXC settings for BT656
- Disable some LCD settings if BT656 is used

Modifications in fsimx6_defconfig:

- Add CONFIG_FB_MXC_BT656 to allow BT656 support
- Add CONFIG_VIDEO_ADV739X to have support for ADV7391