fix usb settings
We have added the flag over-current-active-low. NXP has added now a new
flag which is the opposite of our flag. So we have changed the code to
stay nearly the same as the official NXP kernel. That is why we remove
the flag over-current-active-low. Because if do not set a flag it is
automatically over-current-active-low without setting this bit.