Add more network features to fsimx6/ul/sx defconfigs
The driver for F&S boards using the Silex QCA9377 WLAN/BT chip is closed
source. F&S must provide pre-compiled binaries. However if a customer
changes any options in the network stack this can cause the precompiled
driver to fail. This means we must provide a network stack that has more
features so that most customers can live with it without having to change
anything. Then also the binary driver will work in all these cases.

So add quite a lot of network features to the default network configuration
for fsimx6sx and fsimx6ul. To avoid diverging settings, also add the same
features to fsimx6.

These features add about 250KB to the uImage size of the kernel.