fix compiler errors after merge of several files
esmt_ts.c, ft5x06_ts.c:
The function touchscreen_parse_of_params doesn´t exist anymore. The new
name of the function is called touchscreen_parse_properties. This
function needs some more handover parameters.
adv739x.c:
Some of the functions of the struct v4l2_subdev_core_ops are deleted in
the new kernel. So we delete this calls because in all other drivers
they are also removed.
imx.c:
While we are merged this file, we forget to delete the line. Now the
declaration is moved to the beginnig of the function so we can remove
it.