Use new F&S Tux logo when booting
In the previous logo, Tux held the colorful NetDCU tile in his hands.
This tile is deprecated. Change the tile to the newer version based on
the F&S logo from 2014.

  logo_fs_clut224.ppm    Version with 224 colors
  logo_fs_vga16.ppm      Version with 16 colors (VGA palette)
  logo_fs_mono.pbm       Version with 2 colors (black & white)

Add .png versions of the images, too, as these are smaller and usually
more portable. Also add the .xcf version (for GIMP). This holds an extra
layer for the F&S tile with a mask to hide parts of the tile behind arms
and feet of Tux.

Conversion process
------------------

The clut224 version was created by separating the tile from the original
logo (.svg) in Inkscape and exported to a .png file using a rather large
resolution. Then GIMP was used to scale down the tile to 28x28 pixels.
The result was manually improved so that the rounded edge was better
recognizable. The result was used to replace the tile of the .xcf file
mentioned above. The new .xcf file was saved as full RGB version. After
that GIMP function

  Image -> Mode -> Indexed...

was used to reduce the image to 224 colors and the result was exported
to the .png image. Finally the .png image was converted to .ppm:

  pngtopnm --plain logo_fs_clut224.png > logo_fs_clut224.ppm

The vga16 version is the clut224 version reduced to the VGA color map
(the color map is available as drivers/video/logo/clut_vga16.ppm):

  ppmquant --plain -map clut_vga16.ppm logo_fs_clut224.ppm \
           > logo_fs_vga16.ppm

The mono version is a version edited by hand in GIMP, then converted
to .pbm:

  pngtopam logo_fs_mono.png | pamditherbw | pamtopnm --plain \
           > logo_fs_mono.pbm

Remark:
In the .ppm/.pbm versions, which are plain ASCII, we also added a
comment line to explain the content of the images.