PicoCOMA9X: correct reference to EMMC card detect
Card detect of EMMC is used to determin if EMMC is placed or not.
The correct signal used for that is GPIO2_IO_12.
But cd-gpios property in EMMC node refers to GPIO1_IO_6 which is
incorrect. GPIO1_IO_6 is the card detect signal for SD-Card not EMMC.
The pincontrol was set correctly to GPIO2_IO12