Add support for Toshiba TC358762 DSI to RGB converter
The TC358762 is a de-serializer chip. The serial data stream is supplied
by the baseband prcoessor via the DSI 1.01 interface. TC358762 chip de-
serializes the stream into a parallel one. The parallel output bus can
be either a DPI or a DBI bus. The usage of either DPI or DBI bus is
mutually exclusive.