Improve mipi_dsi_northwest driver - get DT disp params
NXP implemented the mipi_dsi_northwest driver. You can use a converter
which converts MIPI-DSI signal to RGB/HDMI/LVDS signal. NXP added a
display database in file mxc_edid.c. So this is very static and we wont
setup every display in this database and have to recompile the kernel to
support new displays. So we add a way were we can setup displays
dynamically via device tree. The display information is already stored
in  device tree in the lcdif node. So we are able to take this
information instead of values of the display database.
We can now choose if we want to use values from device tree or values
from display database.