Add support for SD1_VSEL on picocoremx6ull
We have the oppurtunity to support sd-card ultra high speed and high
speed mode. For ultra high speed mode the pins will be pulled up to 1.8
V and in high speed mode to 3.3 V. To support these 2 different voltages
we have add a regulator which switch between these 2 voltages. The
driver detects if the sd-card support ultra high speed or high speed
mode and triggers the regulator.