Improve mipi_dsi_northwest driver flag interpretation
We added a statement which compare flag status of the display settings.
Remove this statements, this is not necessary because the
displaysettings should not setup in this case. This is about connection
between the internal lcdif interface and the mipi-dsi interface. Flags
must be defined in a different place e.g. tc35862.