Add CMSPAR option to IMX UART
The termios flag CMSPAR (Mark or Space Parity) can be used to set the
parity bit as a ninth data bit. If CMSPAR is set, the PARODD flag
determines the bit value: If PARODD is also set, the patity bit is set
to 1, if PARODD is not set, the parity bit is set to 0.