Add 12.5pF setting for RTC PCF85063TP, fix two bugs
The RTC chip PCF85063TP has an option to work with crystals with 12.5pF
capacity (default is 7pF). F&S uses crystals of this kind on the boards
PicoCoreMX6SX, PicoCoreMX6UL, PicoCoreMX8MM and efusMX8X, so add a
boolean device tree property cap_12_5pf that switches the CAP_SEL bit in
register CTRL1 if set. For now, only device trees for PicoCoreMX6SX and
PicoCoreMX6UL exist, so add this property there.

The CAP_SEL bit is changed in function pcf85063_dt_init() which is
called in pcf85063_probe() because it will trigger the OS flag. By doing
this, any negative effects can be avoided. For further details, see the
comment in pcf85063_dt_init().

In addition also fix two bugs in this driver.

1. In function pcf85063_stop_clock(), the return value of function
   i2c_smbus_write_byte_data() is just 0 or an error code. But the
   return value ret should be the content of register CTRL1. Change the
   call so that variable ret is not overwritten.

2. In function pcf85063_start_clock(), the STOP bit (bit 5) must be
   cleared when the clock is started. The previous code cleared all bits
   but bit 5. This only worked because of error 1, where the previous
   value of CRTL1 was always zero and thus the STOP bit was already
   cleared. Fix the bitwise AND by using the inverted value.

While we are at it, also add an error message when setting the time if
the year is invalid, i.e. not between 100 and 199 (2000..2099). Before
it was not clear why a date was not accepted.