Handle flexcan names in a common way in device trees
Unfortunately two different schemes to name CAN ports in device trees
have been established on NXP SoCs.

1. Node name: flexcan<n>@<addr>, Label: can<n>
2. Node name: can<n>@<addr>, Label: flexcan<n>

For example imx6qdl.dtsi uses the first type, imx6sx.dtsi uses the
second type. In the past this was not so important, but in kernel
4.9.11, imx6ul.dtsi changed from the second type to the first
type, which is now unfortunately inconsistent to imx6ull.dtsi,
that still uses the second type.

F&S wants to use the same device tree for all their UL boards, whether
they use i.MX6UL or i.MX6ULL. So this inconsistency is most annoying.
When looking at newer device trees for i.MX7D or i.MX8 in different
variants, the second type seems to be used all over the place. So
from our point of view, imx6qdl.dtsi and now imx6ul.dtsi are wrong.
Therefore change these files to the second type to have a common
variant once and forever. This in turn also needs changes in several
board-specific device trees that are also built when calling a
simple make. Even if this affects many files, the modification is
rather simple and should not cause any bigger merge issues in the
future.