Handle USB over-current polarity
Some F&S boards support over-current detection, at least optionally. In
the past F&S had added code to set the polarity for this. Default is active
high, but the user can set this to active low in the device tree by
adding property "over-current-active-low" to the USB node.

In the meantime, NXP has added a similar feature. However they invert
the logic and use "over-current-active-high". This is inconvenient,
because our experience has shown that this is the default. Which means
adding it to the device tree will not achieve anything. However we want
to stay as compatible as possible. Therefore add the detection of both
properties and show an info text what polarity is used by default if
none of them is given in the device tree.

Add either "over-current-active-high" or "over-current-active-low" to
all F&S device trees that support over-current.