Update to newest Atmel MXT touch driver
This is the newest driver from

  https://github.com/atmel-maxtouch/maXTouch_linux.git

It was introduced somewhere between kernel 4.19 and 4.20, but can also
work on older kernels.

Even if the GIT repository contains the whole Linux kernel source code, only
the file drivers/input/touchscreen/atmel_mxt_ts.c is important.

Drop include/linux/platform_data/atmel_mxt_ts.h, the new version does not
support initialization with platform data anymore.

Add property "atmel,cfg_name" to define an own name for the configuration
file. F&S has added a file for mxt224.cfg to the kernel. By doing this, the
driver can also be included in the kernel image, it is not necessary to
build it as kernel module.