Revert RTC PCF8563 driver to previous version
The RTC driver for PCF8563 has a new option to use the CLKOUT pin as a
regular clk source. This means that a consumer client is required to
activate the 32 kHz clock output, otherwise it is switched off.

However on F&S layouts, the 32 kHz clock is obligatory because it is
used as the slow clock for the CPU and also for some WLAN chips, e.g.
on armStoneA9r2. Until we add a more complex modification to the new
driver, go back to the version of kernel 4.1, where the CLKOUT pin is
left in its default state, which is exactly what we need.