Improvements in rev. 1.40 of gar dts
New Version 1.40:

UART4 is added as a new component in the newest revision 1.40 of
gar1 board. The UART4 is a fully connected UART with RX, TX, RTS, CTS,
DTR, DSR, DCD, RI, SHDN, ONLINE signals.

In the previous revision of the board USB_OTG1 power was directly
connected to 5V. Now we are able to toggle the 5V via GPIO.